+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/ReadingInteractor.kt
@@ -130,10 +130,6 @@ class ReadingInteractor(private val readingActivity: ReadingActivity,
                            bookIndex: Int, chapterIndex: Int): List<Verse> =
             bibleReadingManager.readVerses(translationShortName, parallelTranslations, bookIndex, chapterIndex)
 
-    suspend fun readVerse(translationShortName: String, parallelTranslations: List<String>,
-                          verseIndex: VerseIndex): Verse =
-            bibleReadingManager.readVerse(translationShortName, parallelTranslations, verseIndex)
-
     suspend fun readBookNames(translationShortName: String): List<String> =
             bibleReadingManager.readBookNames(translationShortName)